add 'rootwait' to bootargs
Description of 'rootwait' in kernel docs,

    "Wait (indefinitely) for root device to show up.
     Useful for devices that are detected asynchronously
     (e.g. USB and MMC devices)."

The large majority of devices that we integrate use eMMC for storage
medium, and it is quite common to use a USB "live disk" for provisioning.

Without the 'rootwait' argument, mount of rootfs failes occasionally,
it is a simple a race condition at boot and you are only lucky if it works
without it.

Changelog: Title

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>