MEN-3978: Rename environment tools to grub-mender-grubenv-*.
This renames the tools, adjusts the install scripts, and adds a test
for the Makefile installation. There is also a new target,
"install-legacy-tools", which installs the tools using the old names.

Changelog: User space tools, "fw_printenv" and "fw_setenv" have been
renamed to "grub-mender-grubenv-print" and "grub-mender-grubenv-set",
respectively. This has been done in order to avoid conflict with
U-Boot's user space tools, which have the same names.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>