Improve confusing kernel partition selection.
Although it might work, this code section is confusing because
`ptable_type` is set after it is used. I guess the parsing rules allow
for this once it is used again later, but better to be clear.

Switching to use `root` also aligns better with `grub.d` integration,
where setting `root` is the preferred way of locating the kernel.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>