Use right environment location also in standalone mode.
In grub.d mode, `prefix` points to the mender-grub-grubenv folder
already, because that's where the main grub.cfg file is loaded from;
the grub.cfg in the EFI folder is just used to chainload it. But in
standalone mode, the main grub.cfg is in the EFI folder, so we need
something that works relative to that as well. Since the root is the
same in both cases, use that as a basis instead of the `prefix`.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>