fix: Fix grub.d integration which regressed after standalone fix.
In 2ef6724268d024418, we changed the environment path slightly to
become compatible with standalone mode. This path is correct when
using grub.d integration as well, except for one detail: When using
grub.d integration, the script uses two passes (see
`grub.d/README-mender.md`). The path is only correct on the first
pass, because `root` is being adjusted from the boot partition to
the root partition in between those two passes.

The variables are used to load the environment, which we used to do
twice, but it's better to just load it once, and export the relevant
variables instead. So we now load it only during the first pass, and
then we don't use that path at all on the second pass, which both
fixes the issue, simplifies things, and may even make it a tiny bit
faster.

No changelog, since we haven't released the regression.

Changelog: None
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>