fix: Make sure `grub.d` integration works with partition UUIDs.
There are a couple of things that need fixing here.

The main problem is that when using partition UUIDs, the
`mender_kernel_root_base` doesn't exist. This is because partition
UUIDs are independent of storage devices, so there is no logical value
it can have. To fix it, we need to start referring to each partition
individually instead. This requires support from mender-convert, which
needs to start passing in the partition UUID of the boot partition.

Next, since we need to make a distinction between the UUID and
non-UUID case, create a new `grub.d` script which computes the values
centrally, so we don't need to sprinkle conditionals
everywhere. Replace all existing `mender_kernel_root_base` references
with these new references.

With `mender_kernel_root_base` gone, it's now harder to catch
references to unknown partitions in the
`mender_offline_grub-probe_helper` script. We could fix this by doing
individual parsing of each argument and look for unknown partitions,
instead of parsing the whole set of arguments as one line. But that is
a bigger change which I'm not going to do now. Instead, just remove
the check, since it is only a safeguard, and does not affect the
building of images which are already using valid configurations.

I also discovered that the matching expressions could sometimes match
incorrectly because they were not anchored to a word boundary (so
`loop11` could match `loop1` for instance). Added `\b` anchors to
prevent that.

And finally I added a `DEBUG` variable which can be set for more
information about what the `mender_offline_grub-probe_helper` is
doing, which is helpful during debugging.

Changelog: Title
Ticket: MEN-6645

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>