fix: Remove incomplete support for user config GRUB_FONT and GRUB_THEME
The handling of font and theme was left as an unresolved issue at the
time of supporting grub.d integration (SecureBoot for Debian family
OSs) when doing the GRUB integration with `mender-convert`.

This was not made explicit, but the only clue was the prefix `xx_` for
the specific script. Maybe the original author assumed that GRUB would
ignore a non numeric prefix (?).

In any case the integration does not work well in `mender-convert`, and
with SecureBoot enabled the boot script will print warnings as GRUB
tries to access the non-boot partition(s).

When setting `GRUB_FONT`, the `00_header` script by GRUB will probe the
system *at configure time* to identify which is the boot partition and
assign the UUID to a GRUB search command. In the context of
`mender-convert`, this identification will be erroneous because it will
find the boot partition of the image being converted (the sdimg file
that you would flash to a physical device) and not the boot partition of
the device on the field (the one being updated with a Mender Artifact).

This is the culprit line:
* https://git.savannah.gnu.org/cgit/grub.git/tree/util/grub.d/00_header.in?h=grub-2.12#n150

By removing the file we make it explicit that it is
not supported: Mender GRUB integration only works with default GRUB
font, theme and background.

We will be working together with upstream to find a way to support these
in the future.

Ticket: MEN-8036
Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>